Add calculateTotalPrice

+++ src/main/java/org/example/ProductRepo/Product.java
@@ -1,6 +1,8 @@
 package org.example.ProductRepo;
 
-public record Product(int id, String name, double price) {
+import java.math.BigDecimal;
+
+public record Product(int id, String name, BigDecimal price) {
     @Override
     public String toString() {
         return String.format("{id=%d, name='%s', price=%.2f}", id, name, price);
